deleted LoginDatabaseController, moved LoginDatabaseGateway to database folder, changed submitForm and validateForm to protected in RegisterForm
