+++ src/main/java/database/DatabaseController.java
@@ -1,6 +1,12 @@
 package database;
 
-public class DatabaseController<T> implements CreateListingDatabaseGateway, ReviewDatabaseGateway{
+import entities.User;
+
+public class DatabaseController<T> implements CreateListingDatabaseGateway, ReviewDatabaseGateway, LoginDatabaseGateway{
     String table = null;
 
+    @Override
+    public User getUserWithUsername(String username) {
+        return null;
+    }
 }
