+++ src/main/java/forms/RegisterForm.java
@@ -12,7 +12,7 @@ public class RegisterForm extends Form {
     }
 
     @Override
-    boolean validateForm() {
+    protected boolean validateForm() {
         //password and confirm password
         if (!((this.password.equals(this.confirmPassword)) && (password.length() >= 8))) {
             //output error msg
@@ -32,7 +32,7 @@ public class RegisterForm extends Form {
     }
 
     @Override
-    void submitForm() {
+    protected void submitForm() {
 
     }
